html/browsers/browsing-the-web/navigating-across-documents/012-async-timeout.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<title>Link with onclick, which executes a handler containing a timeout, that then navigates the iframe, and returns true.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="(function(){setTimeout(()=>document.querySelector('iframe').src='call-parent-postmessage.html', 500);reports.push('onclick');})()" href="href.html">Test</a>
<script>
var test = async_test();
var reports = [];

window.onmessage = test.step_func(function(e) {
  reports.push(e.data);

  // "call-parent-postmessage" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "call-parent-postmessage") {
    assert_equals(reports[0], "click", reports);
    assert_equals(reports[1], "onclick", reports);
    assert_equals(reports[2], "href", reports);
    assert_equals(reports[3], "call-parent-postmessage");
    assert_equals(reports.length, 4);
    test.done();
  }
});

test.step(function() {
  reports.push("click");
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates an iframe and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to navigate the iframe after an async timeout.
3. Navigating the iframe to the anchor's `href` will complete successfully,
    then the queued timeout will execute and the next navigation will occur
    and also complete successfully.
4. If the semantics are implemented correctly, reports will be
    ["click", "onclick", "href", "call-parent-postmessage"].

-->
